+++ Frends.Hl7v2.CreateFromXml/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2026-07-03
+
+### Added
+
+- MSH field overrides (MshOverrides) — ability to overwrite any MSH field (MSH-1 to MSH-15) in the output HL7v2 message
+CR option to LineEnding enum
+
 ## [1.0.0] - 2026-02-27
 
 ### Added